Refactor : Game.jsx style과 로직 분리
상세 : Game.jsx에서 논리를 담당하는 index.js와
스타일을 담당하는 style.js 파일로 파일 분리
이유: 로직과 스타일을 분리해서 관리하는 것이 좋을 것이라 판단
